DAY AFTER SUBMISSION

+++ src/boundary.rs
@@ -1,15 +1,15 @@
 #[derive()]
 pub struct Boundary {
-    rangex: Vec<f32>,
-    rangey: Vec<f32>,
+    rangex: Vec<f64>,
+    rangey: Vec<f64>,
 }
 
 impl Boundary {
-    pub fn get_xrange(&self) -> &Vec<f32> {
+    pub fn get_xrange(&self) -> &Vec<f64> {
         &self.rangex
     }
 
-    pub fn get_yrange(&self) -> &Vec<f32> {
+    pub fn get_yrange(&self) -> &Vec<f64> {
         &self.rangey
     }
 }
